tentative of refacto - not functioning
